AI_UseCase/requirements.txt
streamlit
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
sentence-transformers
faiss-cpu
pymupdf
tavily-python
numpy
python-dotenv